requirements.txt
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorclass==2.2.2
compressed_rtf==1.0.6
cryptography==43.0.0
dj-database-url==2.2.0
Django==5.0.7
django-allauth==64.0.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
easygui==0.98.3
ebcdic==1.1.1
extract-msg==0.48.7
gunicorn==22.0.0
h11==0.14.0
idna==3.7
lark==1.1.9
msoffcrypto-tool==5.4.1
oauthlib==3.2.2
olefile==0.47
oletools==0.60.2
packaging==24.1
pcodedmp==1.2.6
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.2
python-dotenv==1.0.1
red-black-tree-mod==1.20
requests==2.32.3
requests-oauthlib==2.0.0
RTFDE==0.1.2
soupsieve==2.5
sqlparse==0.5.1
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.5
whitenoise==6.7.0
